requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
executing==1.0.0
fastapi==0.95.0
fastjsonschema==2.16.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.14.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
openai==0.27.2
orjson==3.8.9
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
rfc3986==1.5.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.26.1
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.12
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
widgetsnbextension==4.0.2
yarl==1.8.2
zipp==3.8.1
